Refactor upgrades
